Put streaming tests into own file
